+++ src/libsystemd-network/fuzz-ndisc-rs.c
@@ -11,7 +11,7 @@
 #include "fd-util.h"
 #include "fuzz.h"
 #include "icmp6-packet.h"
-#include "icmp6-util-unix.h"
+#include "icmp6-test-util.h"
 #include "ndisc-internal.h"
 #include "ndisc-option.h"
 #include "socket-util.h"
